Move Operation to its own module
