+++ tests/unit/converter/test_organize_rows.py
@@ -8,7 +8,8 @@ from nd2k.converter.organize_rows import (
 	create_operation,
 	create_partial_trade,
 )
-from nd2k.types import OperationType, PartialTrade, TradingPair, Operation
+from nd2k.types import PartialTrade, TradingPair
+from nd2k.operation import OperationType, Operation
 from ..helpers import fake_op
