+++ tests/unit/helpers.py
@@ -1,12 +1,8 @@
 from typing import Any
 from datetime import datetime
 from decimal import Decimal
-from nd2k.types import (
-	Operation,
-	OperationType,
-	PartialTrade,
-	TradingPair,
-)
+from nd2k.types import PartialTrade, TradingPair
+from nd2k.operation import OperationType, Operation
 
 
 def fake_op(**kwargs: Any) -> Operation:
